+++ data/admin/abstract_admin.py
@@ -1,8 +1,8 @@
 from simple_history.admin import SimpleHistoryAdmin
 
 
-class IngredientAdminWithHistoryChangedFields(SimpleHistoryAdmin):
-    history_list_display = ['changed_fields']
+class IngredientAdminHistorisableChangedFields(SimpleHistoryAdmin):
+    history_list_display = ["changed_fields"]
 
     def changed_fields(self, obj):
         if obj.prev_record:
