+++ data/behaviours/__init__.py
@@ -1,3 +1,4 @@
 from .auto_validable import AutoValidable  # noqa: F401
 from .time_stampable import TimeStampable  # noqa: F401
+from .historisable import Historisable  # noqa: F401
 from .deactivable import Deactivable, DeactivableQuerySet  # noqa: F401
