+++ data/behaviours/historisable.py
@@ -1,0 +1,9 @@
+from django.db import models
+from simple_history.models import HistoricalRecords
+
+
+class Historisable(models.Model):
+    class Meta:
+        abstract = True
+
+    history = HistoricalRecords(inherit=True)
