+++ data/models/microorganism.py
@@ -4,7 +4,8 @@ from django.db.models import F, Value
 
 from simple_history.models import HistoricalRecords
 
-from .mixins import WithCreationAndModificationDate, WithHistory, WithMissingImportBoolean, WithComments
+from data.behaviours import TimeStampable, Historisable
+from .mixins import WithMissingImportBoolean, WithComments
 from .abstract_models import CommonModel
 from .substance import Substance
 
@@ -29,7 +30,7 @@ class Microorganism(CommonModel, WithComments):
         return self.siccrf_name_en
 
 
-class MicroorganismSubstanceRelation(WithCreationAndModificationDate, WithHistory):
+class MicroorganismSubstanceRelation(TimeStampable, Historisable):
     microorganism = models.ForeignKey(Microorganism, on_delete=models.CASCADE)
     substance = models.ForeignKey(Substance, on_delete=models.CASCADE)
     siccrf_is_related = models.BooleanField(
@@ -38,7 +39,7 @@ class MicroorganismSubstanceRelation(WithCreationAndModificationDate, WithHistor
     ca_is_related = models.BooleanField(null=True, default=None, verbose_name="substance associée au micro-organisme")
 
 
-class MicroorganismSynonym(WithCreationAndModificationDate, WithHistory, WithMissingImportBoolean):
+class MicroorganismSynonym(TimeStampable, Historisable, WithMissingImportBoolean):
     class Meta:
         verbose_name = "synonyme de micro-organisme"
