+++ data/models/plant.py
@@ -4,7 +4,8 @@ from django.db.models import F
 
 from simple_history.models import HistoricalRecords
 
-from .mixins import WithCreationAndModificationDate, WithHistory, WithMissingImportBoolean, WithComments
+from data.behaviours import TimeStampable, Historisable
+from .mixins import WithMissingImportBoolean, WithComments
 from .abstract_models import CommonModel
 from .substance import Substance
 
@@ -63,7 +64,7 @@ class Plant(CommonModel, WithComments):
     history = HistoricalRecords(inherit=True, excluded_fields=["name", "is_obsolete", "family"])
 
 
-class Part(WithCreationAndModificationDate):
+class Part(TimeStampable):
     """
     Ce modèle permet d'associer des données supplémentaires à la relation ManyToMany
     plant_parts
@@ -92,7 +93,7 @@ class Part(WithCreationAndModificationDate):
     )
 
 
-class PlantSubstanceRelation(WithCreationAndModificationDate, WithHistory):
+class PlantSubstanceRelation(TimeStampable, Historisable):
     plant = models.ForeignKey(Plant, on_delete=models.CASCADE)
     substance = models.ForeignKey(Substance, on_delete=models.CASCADE)
     siccrf_is_related = models.BooleanField(
@@ -101,7 +102,7 @@ class PlantSubstanceRelation(WithCreationAndModificationDate, WithHistory):
     ca_is_related = models.BooleanField(null=True, default=None, verbose_name="substance associée à la plante")
 
 
-class PlantSynonym(WithCreationAndModificationDate, WithHistory, WithMissingImportBoolean):
+class PlantSynonym(TimeStampable, Historisable, WithMissingImportBoolean):
     class Meta:
         verbose_name = "synonyme de plante"
